// personal-finance-manager-frontend/app/src/app/models/categorias.enum.ts
export enum TipoCategoria {
  COMIDASYBEBIDAS = 'COMIDASYBEBIDAS',
  SUPERMERCADO = 'SUPERMERCADO',
  ENTRENAMIENTO = 'ENTRENAMIENTO',
  TRANSPORTE = 'TRANSPORTE',
  AMISTADESYFAMILIA = 'AMISTADESYFAMILIA',
  CREDITOSYFINANCIACION = 'CREDITOSYFINANCIACION',
  CUENTASYSERVICIOS = 'CUENTASYSERVICIOS',
  EDUCACION = 'EDUCACION',
  HOGAR = 'HOGAR',
  IMPUESTOS = 'IMPUESTOS',
  INDUMENTARIA = 'INDUMENTARIA',
  INVERSIONES = 'INVERSIONES',
  MASCOTAS = 'MASCOTAS',
  OTROS = 'OTROS',
  SALUDYCUIDADOPERSONAL = 'SALUDYCUIDADOPERSONAL',
  SHOPPING = 'SHOPPING',
  SUSCRIPCIONES = 'SUSCRIPCIONES',
  VIAJES = 'VIAJES',
  PENDIENTEDECATEGORIA = 'PENDIENTEDECATEGORIA',
  SUELDO = 'SUELDO'
}